AppUserMode.cpp: Set true and false in ther proper place
Moved Portable file to pcsx2/Docs. This prevents it being modified when it's set to false. Ignore bin portable file in git. VS: Post-build event to copy portable.yaml file into /bin with escape-hatch

If a /bin/portable.build.yaml file is detected, it will use that instead of the pcsx2/Docs/portable.yaml file.

This is to maintain parity with the CMake solution which is to always overwrite the file while still allowing a decent method to bypass that behaviour.